+++ src/backend/bench/kpro.py
@@ -8,11 +8,15 @@ from backend.devices.kpro.kpro import Kpro
 kpro = Kpro()
 with output(output_type="dict", initial_len=1, interval=0) as output_list:
     while True:
+        output_list["VTS"] = str(kpro.vts)
+        output_list["VTP"] = str(kpro.vtp)
+        output_list["VTEC"] = str(kpro.vtec)
         output_list["ETH"] = str(kpro.eth)
         output_list["FLT"] = str(kpro.flt)
         output_list["BAT"] = str(kpro.bat)
         output_list["CAM"] = str(kpro.cam)
         output_list["O2"] = str(kpro.o2)
+        output_list["O2_CMD"] = str(kpro.o2_cmd)
         output_list["IAT"] = str(kpro.iat)
         output_list["RPM"] = str(kpro.rpm)
         output_list["TPS"] = str(kpro.tps)