+++ src/backend/websocket.py
@@ -4,6 +4,8 @@ import threading
 
 import websockets
 
+from backend.constants import WEBSOCKET_HOST, WEBSOCKET_PORT
+from backend.devices.logger.logger import Logger
 from backend.devices.setup_validator.setup_validator import SetupValidator
 
 
@@ -13,7 +15,9 @@ class Websocket:
 
     def __init__(self, backend):
         self.backend = backend
-        self.websocket = websockets.serve(self._websocket_handler, "0.0.0.0", 5678)
+        self.websocket = websockets.serve(
+            self._websocket_handler, WEBSOCKET_HOST, WEBSOCKET_PORT
+        )
         self.loop = asyncio.get_event_loop()
         self.loop.run_until_complete(self.websocket)
         threading.Thread(target=self.loop.run_forever).start()
@@ -34,6 +38,23 @@ class Websocket:
             data = json.loads(message)
             if data["action"] == "setup":
                 await websocket.send(json.dumps({"setup": self.backend.setup()}))
+            elif data["action"] == "toggle_style":
+                self.backend.style.toggle()
+            elif data["action"] == "toggle_datalog":
+                self.backend.logger.toggle()
+            elif data["action"] == "datalogs":
+                await websocket.send(json.dumps({"datalogs": Logger.get_logs()}))
+            elif data["action"].startswith("datalog_"):
+                file_name = data["action"].split("datalog_")[1]
+                log = Logger.get_log(file_name)
+                if log:
+                    await websocket.send(json.dumps({f"datalog_{file_name}": log}))
+            elif data["action"].startswith("remove_datalog_"):
+                try:
+                    Logger.remove_log(data["action"].split("remove_datalog_")[1])
+                except FileNotFoundError:
+                    pass
+                await websocket.send(json.dumps({"datalogs": Logger.get_logs()}))
             elif data["action"] == "save":
                 try:
                     self.backend.save(data["data"])
@@ -58,11 +79,27 @@ class Websocket:
                 await self._send_all_clients(
                     json.dumps({"action": "alert", "message": "SUCCESS: setup reset!"})
                 )
+            elif data["action"] == "raw":
+                await websocket.send(
+                    json.dumps(
+                        {
+                            "data0": list(getattr(self.backend.ecu.ecu, "data0", [])),
+                            "data1": list(getattr(self.backend.ecu.ecu, "data1", [])),
+                            "data2": list(getattr(self.backend.ecu.ecu, "data2", [])),
+                            "data3": list(getattr(self.backend.ecu.ecu, "data3", [])),
+                            "data4": list(getattr(self.backend.ecu.ecu, "data4", [])),
+                            "data5": list(getattr(self.backend.ecu.ecu, "data5", [])),
+                            "data6": list(getattr(self.backend.ecu.ecu, "data6", [])),
+                        }
+                    )
+                )
 
     async def _producer_handler(self, websocket):
         """Keeps sending updated ecu data forever"""
         while True:
-            await websocket.send(json.dumps({"data": self.backend.update()}))
+            data = self.backend.update()
+            await websocket.send(json.dumps({"data": data}))
+            self.backend.logger.log(data)
             await asyncio.sleep(0.1)
 
     async def _send_all_clients(self, message):