+++ src/tests/backend/devices/setup_updater/test_setup_updater.py
@@ -61,7 +61,28 @@ class TestSetupUpdater:
         assert m__update_to_3_2_0.called is True
 
     def test_update_3_2_0(self):
-        assert SetupUpdater().update({"version": "3.2.0"}) == {"version": "3.2.0"}
+        with mock.patch(
+            "backend.devices.setup_updater.SetupUpdater._update_to_3_3_0"
+        ) as m__update_to_3_3_0:
+            SetupUpdater().update({"version": "3.2.0"})
+        assert m__update_to_3_3_0.called is True
+
+    def test_update_3_3_0(self):
+        with mock.patch(
+            "backend.devices.setup_updater.SetupUpdater._update_to_3_4_0"
+        ) as m__update_to_3_4_0:
+            SetupUpdater().update({"version": "3.3.0"})
+        assert m__update_to_3_4_0.called is True
+
+    def test_update_3_4_0(self):
+        with mock.patch(
+            "backend.devices.setup_updater.SetupUpdater._update_to_3_5_0"
+        ) as m__update_to_3_5_0:
+            SetupUpdater().update({"version": "3.4.0"})
+        assert m__update_to_3_5_0.called is True
+
+    def test_update_3_5_0(self):
+        assert SetupUpdater().update({"version": "3.5.0"}) == {"version": "3.5.0"}
 
     def test__update_to_2_4_0(self, fixtures_dir):
         with open(os.path.join(fixtures_dir, "default_setup_2_3_2.json")) as file:
@@ -110,6 +131,30 @@ class TestSetupUpdater:
         with open(os.path.join(fixtures_dir, "default_setup_3_2_0.json")) as file:
             fixture_3_2_0 = json.load(file)
         assert setup == fixture_3_2_0
+
+    def test__update_to_3_3_0(self, fixtures_dir):
+        with open(os.path.join(fixtures_dir, "default_setup_3_2_0.json")) as file:
+            setup = json.load(file)
+        setup = SetupUpdater._update_to_3_3_0(setup)
+        with open(os.path.join(fixtures_dir, "default_setup_3_3_0.json")) as file:
+            fixture_3_3_0 = json.load(file)
+        assert setup == fixture_3_3_0
+
+    def test__update_to_3_4_0(self, fixtures_dir):
+        with open(os.path.join(fixtures_dir, "default_setup_3_3_0.json")) as file:
+            setup = json.load(file)
+        setup = SetupUpdater._update_to_3_4_0(setup)
+        with open(os.path.join(fixtures_dir, "default_setup_3_4_0.json")) as file:
+            fixture_3_4_0 = json.load(file)
+        assert setup == fixture_3_4_0
+
+    def test__update_to_3_5_0(self, fixtures_dir):
+        with open(os.path.join(fixtures_dir, "default_setup_3_4_0.json")) as file:
+            setup = json.load(file)
+        setup = SetupUpdater._update_to_3_5_0(setup)
+        with open(os.path.join(fixtures_dir, "default_setup_3_5_0.json")) as file:
+            fixture_3_5_0 = json.load(file)
+        assert setup == fixture_3_5_0
         SetupValidator().validate(
             setup
         )  # this only could be applied to the last version update test