+++ CHANGELOG.md
@@ -4,6 +4,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.5.0] - 2022-05-28
+### Added
+- Data logger
+- Screen toggle
+- Debug screens
+- Target O2 value (untested)
+
+### Changed
+- Upgraded to python 3.9
+- Raspberry Pi OS 64 bit base image
+- Alternative layout fixes
+
+## [3.4.0] - 2021-06-28
+### Added
+- VTEC for K-Pro v2/3/4 and S300 v3
+- FANC for S300 v3
+
+### Changed
+- Fix SCS for S300 v3
+
+## [3.3.0] - 2021-06-16
+### Added
+- Analog inputs for S300 v3
+
 ## [3.2.0] - 2021-03-23
 ### Changed
 - Fixed specific pressure and temperature sensor formulas breaking the app when selected